Make log dir paths with mkpath
The directory creation code has been replaced with a call to mkpath.  The
previous code has a race condition where it checks for the directory's
existence, then emits an error if mkdir fails (but it's entirely possible
another process can create the directory inbetween these calls, as I've had it
happen).  Instead, mkpath ensures the directories exist and only emits errors
if they cannot be created.

File::Path has been included in Perl for a long time (at least as far back as
5.8.8), so this shouldn't be adding a dependency.