Add date search option to node.html
Add another search parameter to node.html that restricts the results to entries
that have the given date within their timespan.  For example, this can be used
to find what MAC addresses could have been using a given IP address at a
specific time, instead of forcing the user to do this manually by comparing all
of the dates themselves.