Modify columns on device_inv vlan summary
Previously, the vlan summary had a column labeled "Count" which was rather
ambigious.  It ended up being a count of the number of rows device_vlan has
with that particular vlan/name pair, which is basically the count of devices
that have the vlan configured in their vlan database (at least, as far as Cisco
devices behave).  This means that, if VTP is used or vlans are otherwise named
on devices where there are no active ports, the count exaggerates how many
devices actually service the vlan.

Instead, this patch removes the "Count" column and adds "Device Count" and
"Port Count" columns.  The "Device Count" column, instead of holding the number
of devices with that vlan/name pair, is the number of devices with interfaces
configured with that vlan.  The "Port Count" column contains the number of
interfaces with that vlan configured on them.

Ideally, the information presented by the "Count" column would still be
available somehow but I was unable to get such results within one query (it is
beyond the limits of my SQL skills).  However, it's hard to imagine a scenario
where a count of the number of devices with a vlan configured but not in use
would be helpful.  The user is more likely expecting the number that "Device
Count" provides.