Correct error message during discovery
One of our devices was experiencing packet loss and causing the discovery
process to try to print $layers when it was undefined (presumably because the
SNMP queries to retrieve it were failing).  Since SNMP::Info supports devices
with strange and/or non-existent $layers, this error message seems out of date.
Instead of reporting $layers, it seems more accurate to report that the device
simply isn't supported by SNMP::Info.  The error message has been changed to
reflect this.