imported wiki pages
The SF wiki is going to be removed. We only have a small number of pages
there so I copied the content into some POD files. Some of it may be out
of date, so they can be deleted but at least there's now a backup (I'm
going to kill the wiki, now).