Fix infinite loop when an early worker dies but a later worker remains
This case was very nearly handled properly.  If we are worker 4,
meaning that we are using offset 4, and there are less than 4 jobs
remaining in the queue, we set $offset to 0.  Unfortunately, we
then loop back to the top and re-set $offset to $i (4).  The fix
is to move the initialization of $offset outside the loop, so that
the update to 0 can stick.

The symptom of this bug is that there is one worker hanging
around hammering on the database.  On my system it was one
netdisco process and one postmaster process sharing a cpu 50/50.
It only happens in unusual situations where earlier workers
have died unexpectedly (in my case apparently because of
running into database connection limits).